Interactive mode + fix eventset's schema
