+++ pyproject.toml
@@ -28,6 +28,7 @@ mkdocs-material = "^9.1.5"
 mkdocs-gen-files = "^0.4.0"
 mkdocs-literate-nav = "^0.6.0"
 mkdocs-section-index = "^0.3.5"
+bokeh = "^3.1.0"
 
 [tool.black]
 line-length = 80
