+++ temporian/core/serialize.py
@@ -173,6 +173,8 @@ def unserialize(src: pb.Processor) -> processor.Preprocessor:
 
 def _identifier(item: Any) -> str:
     """Creates a unique identifier for an object within a processor."""
+    if item is None:
+        raise ValueError("Cannot get id of None")
     return str(id(item))
 
 
@@ -262,7 +264,7 @@ def _unserialize_event(
     src: pb.Event, samplings: Dict[str, Sampling], features: Dict[str, Feature]
 ) -> Event:
     if src.sampling_id not in samplings:
-        raise ValueError(f"Non existing sampling {src.sampling_id}")
+        raise ValueError(f"Non existing sampling {src.sampling_id} in {src}")
 
     def get_feature(key):
         if key not in features:
@@ -293,7 +295,7 @@ def _unserialize_feature(
     src: pb.Feature, samplings: Dict[str, Sampling]
 ) -> Feature:
     if src.sampling_id not in samplings:
-        raise ValueError(f"Non existing sampling {src.sampling_id}")
+        raise ValueError(f"Non existing sampling {src.sampling_id} in {src}")
 
     return Feature(
         name=src.name,
@@ -308,7 +310,7 @@ def _serialize_sampling(src: Sampling) -> pb.Sampling:
         id=_identifier(src),
         index=pb.Index(
             levels=[
-                pb.Index.IndexLevel(name, dtype)
+                pb.Index.IndexLevel(name=name, dtype=_serialize_dtype(dtype))
                 for name, dtype in zip(src.index.names, src.index.dtypes)
             ]
         ),
@@ -321,7 +323,7 @@ def _serialize_sampling(src: Sampling) -> pb.Sampling:
 def _unserialize_sampling(src: pb.Sampling) -> Sampling:
     return Sampling(
         index_levels=[
-            (index_level.name, index_level.dtype)
+            (index_level.name, _unserialize_dtype(index_level.dtype))
             for index_level in src.index.levels
         ],
         creator=None,
