+++ temporian/core/test/BUILD
@@ -59,6 +59,7 @@ py_test(
         # absl/logging dep,
         # absl/testing:absltest dep,
         ":utils",
+        "//temporian",
         "//temporian/core:processor",
         "//temporian/core:serialize",
         "//temporian/core/data:dtype",
