+++ temporian/core/test/utils.py
@@ -23,7 +23,8 @@ def create_input_event():
         features=[
             Feature("f1", DType.FLOAT32),
             Feature("f2", DType.FLOAT32),
-        ]
+        ],
+        index_levels=[("x", DType.INT32), ("y", DType.STRING)],
     )
