+++ temporian/implementation/numpy/data/plotter.py
@@ -1,5 +1,6 @@
-import datetime
-from typing import NamedTuple, Optional, Union, List, Any, Tuple
+"""Plotting utility."""
+
+from typing import NamedTuple, Optional, Union, List, Any
 
 import numpy as np
 from enum import Enum
@@ -7,22 +8,24 @@ from enum import Enum
 from temporian.core.data import duration
 from temporian.implementation.numpy.data.event import NumpyEvent
 
-DEFAULT_BACKEND = "bokeh"
-
 
 class Style(Enum):
     """Plotting style."""
 
+    # Determine the style according to the data.
     auto = "auto"
+    # Connect numerical values with a line.
     line = "line"
+    # A discreet marker showing a feature value.
     marker = "marker"
+    # A discreet marker not showing a feature value.
     vline = "vline"
 
 
 class Options(NamedTuple):
     """Options for plotting."""
 
-    backend: str
+    backend: Optional[str]
     height_per_plot_px: int
     width_px: int
     max_points: Optional[int]
@@ -30,12 +33,12 @@ class Options(NamedTuple):
     max_time: Optional[duration.Timestamp]
     max_num_plots: int
     style: Style
+    interactive: bool
 
 
 def plot(
     events: Union[List[NumpyEvent], NumpyEvent],
     indexes: Optional[Union[Any, tuple, List[tuple]]] = None,
-    backend: str = DEFAULT_BACKEND,
     width_px: int = 1024,
     height_per_plot_px: int = 150,
     max_points: Optional[int] = None,
@@ -44,6 +47,8 @@ def plot(
     max_num_plots: int = 20,
     style: Union[Style, str] = Style.auto,
     return_fig: bool = False,
+    interactive: bool = False,
+    backend: Optional[str] = None,
 ):
     """Plots an event.
 
@@ -53,7 +58,6 @@ def plot(
             the available indexes. Indexes should be provided as single value
             (e.g. string) or tuple of values. Example: index="a", index=("a",),
             index=("a", "b",), index=["a", "b"], index=[("a", "b"), ("a", "c")].
-        backend: Plotting library to use.
         width_px: Width of the figure in pixel.
         height_per_plot_px: Height of each sub-plot (one per feature) in pixel.
         max_points: Maximum number of points to plot.
@@ -64,6 +68,10 @@ def plot(
             warning.
         return_fig: If true, returns the figure object. The figure object
             depends on the backend.
+        interactive: If true, creates an interractive plotting. interactive=True
+            effectively selects a backend that support interractive plotting.
+        backend: Plotting library to use. Possible values are: matplotlib,
+            bokeh.
     """
 
     original_indexes = indexes
@@ -102,6 +110,7 @@ def plot(
     assert isinstance(style, Style)
 
     options = Options(
+        interactive=interactive,
         backend=backend,
         width_px=width_px,
         height_per_plot_px=height_per_plot_px,
@@ -112,20 +121,29 @@ def plot(
         style=style,
     )
 
+    if backend is None:
+        backend = "bokeh" if interactive else "matplotlib"
+
     if backend not in BACKENDS:
         raise ValueError(
             f"Unknown plotting backend {backend}. Available "
             f"backends: {BACKENDS}"
         )
 
-    fig = BACKENDS[backend](events=events, indexes=indexes, options=options)
+    try:
+        fig = BACKENDS[backend](events=events, indexes=indexes, options=options)
+    except ImportError:
+        print(error_message_import_backend(backend))
+        raise
+
     return fig if return_fig else None
 
 
-def _num_plots(
+def get_num_plots(
     events: List[NumpyEvent], indexes: List[tuple], options: Options
 ):
-    """Compute the number of sub-plots + extra checks."""
+    """Computes the number of sub-plots."""
+
     num_plots = 0
     for index in indexes:
         for event in events:
@@ -157,185 +175,6 @@ def _num_plots(
     return num_plots
 
 
-def _plot_matplotlib(
-    events: List[NumpyEvent], indexes: List[tuple], options: Options
-):
-    import matplotlib.pyplot as plt
-    from matplotlib.cm import get_cmap
-
-    colors = get_cmap("tab10").colors
-
-    px = 1 / plt.rcParams["figure.dpi"]
-
-    num_plots = _num_plots(events, indexes, options)
-
-    fig, axs = plt.subplots(
-        num_plots,
-        figsize=(
-            options.width_px * px,
-            options.height_per_plot_px * num_plots * px,
-        ),
-        squeeze=False,
-    )
-
-    # Actual plotting
-    plot_idx = 0
-    for index in indexes:
-        if plot_idx >= num_plots:
-            # Too much plots are displayed already.
-            break
-
-        # Note: Don't display the tuple parenthesis is the index contain a
-        # single dimension.
-        title = str(index[0] if len(index) == 1 else index)
-
-        # Index of the next color to use in the plot.
-        color_idx = 0
-
-        for event in events:
-            if plot_idx >= num_plots:
-                break
-
-            feature_names = event.feature_names
-
-            xs = event.data[index].timestamps
-            uniform = is_uniform(xs)
-
-            plot_mask = np.full(len(xs), True)
-            if options.min_time is not None:
-                plot_mask = plot_mask & (xs >= options.min_time)
-            if options.max_time is not None:
-                plot_mask = plot_mask & (xs <= options.max_time)
-            if options.max_points is not None and len(xs) > options.max_points:
-                # Too many timestamps. Only keep the fist ones.
-                plot_mask = plot_mask & (
-                    np.cumsum(plot_mask) <= options.max_points
-                )
-
-            xs = xs[plot_mask]
-
-            if event.is_unix_timestamp:
-                # Matplotlib understands datetimes.
-                xs = [
-                    datetime.datetime.fromtimestamp(x, tz=datetime.timezone.utc)
-                    for x in xs
-                ]
-
-            if len(feature_names) == 0:
-                # There is not features to plot. Instead, plot the timestamps.
-                _matplotlib_sub_plot(
-                    ax=axs[plot_idx, 0],
-                    xs=xs,
-                    ys=np.zeros(len(xs)),
-                    options=options,
-                    color=colors[color_idx % len(colors)],
-                    name="[sampling]",
-                    is_unix_timestamp=event.is_unix_timestamp,
-                    title=title,
-                    style=Style.vline,
-                )
-                # Only print the index / title once
-                title = None
-
-                color_idx += 1
-                plot_idx += 1
-
-            for feature_idx, feature_name in enumerate(feature_names):
-                if plot_idx >= num_plots:
-                    # Too much plots are displayed already.
-                    break
-
-                ys = event.data[index].features[feature_idx][plot_mask]
-                if len(ys) == 0:
-                    all_ys_are_equal = True
-                else:
-                    all_ys_are_equal = np.all(ys == ys[0])
-
-                effective_stype = options.style
-                if effective_stype == Style.auto:
-                    if not uniform and (len(xs) <= 1000 or all_ys_are_equal):
-                        effective_stype = Style.marker
-                    else:
-                        effective_stype = Style.line
-
-                _matplotlib_sub_plot(
-                    ax=axs[plot_idx, 0],
-                    xs=xs,
-                    ys=ys,
-                    options=options,
-                    color=colors[color_idx % len(colors)],
-                    name=feature_name,
-                    is_unix_timestamp=event.is_unix_timestamp,
-                    title=title,
-                    style=effective_stype,
-                )
-
-                # Only print the index / title once
-                title = None
-
-                color_idx += 1
-                plot_idx += 1
-
-    fig.tight_layout()
-    return fig
-
-
-def _matplotlib_sub_plot(
-    ax,
-    xs,
-    ys,
-    options: Options,
-    color,
-    name: str,
-    is_unix_timestamp: bool,
-    title: Optional[str],
-    style: Style,
-    **wargs,
-):
-    """Plots "(xs, ys)" on the axis "ax"."""
-
-    import matplotlib.ticker as ticker
-
-    if style == Style.line:
-        mat_style = {}  # Default
-    elif style == Style.marker:
-        mat_style = {"marker": "2", "linestyle": "None"}
-    elif style == Style.vline:
-        mat_style = {"marker": "|", "linestyle": "None"}
-    else:
-        raise ValueError("Non implemented style")
-
-    ax.plot(xs, ys, lw=0.5, color=color, **mat_style, **wargs)
-    if options.min_time is not None or options.max_time is not None:
-        args = {}
-        if options.min_time is not None:
-            args["left"] = (
-                duration.convert_date_to_duration(options.min_time)
-                if not is_unix_timestamp
-                else options.min_time
-            )
-        if options.max_time is not None:
-            args["right"] = (
-                duration.convert_date_to_duration(options.max_time)
-                if not is_unix_timestamp
-                else options.max_time
-            )
-        ax.set_xlim(**args)
-
-    ax.xaxis.set_tick_params(labelsize=8)
-    ax.xaxis.set_major_locator(ticker.MaxNLocator(10))
-    ax.xaxis.set_minor_locator(ticker.NullLocator())
-
-    ax.set_ylabel(name, size=8)
-    ax.yaxis.set_tick_params(labelsize=8)
-    ax.yaxis.set_major_locator(ticker.MaxNLocator(5))
-    ax.yaxis.set_minor_locator(ticker.NullLocator())
-
-    ax.grid(lw=0.4, ls="--", axis="x")
-    if title:
-        ax.set_title(title)
-
-
 def is_uniform(xs) -> bool:
     """Checks if timestamps are uniformly sampled."""
 
@@ -345,245 +184,18 @@ def is_uniform(xs) -> bool:
     return np.allclose(diff, diff[0])
 
 
-def _plot_bokeh(
-    events: List[NumpyEvent], indexes: List[tuple], options: Options
-):
-    from bokeh.plotting import figure
-    from bokeh.io import output_notebook, show
-    from bokeh.layouts import gridplot, column
-    from bokeh.models import ColumnDataSource, CategoricalColorMapper, HoverTool
-    from bokeh.models import ColumnDataSource, CustomJS
-    from bokeh.palettes import Dark2_5 as colors
+from temporian.implementation.numpy.data.plotter_bokeh import plot_bokeh
+from temporian.implementation.numpy.data.plotter_matplotlib import (
+    plot_matplotlib,
+)
 
-    num_plots = _num_plots(events, indexes, options)
-
-    figs = []
-
-    # Actual plotting
-    plot_idx = 0
-    for index in indexes:
-        if plot_idx >= num_plots:
-            # Too much plots are displayed already.
-            break
-
-        # Note: Don't display the tuple parenthesis is the index contain a
-        # single dimension.
-        title = str(index[0] if len(index) == 1 else index)
-
-        # Index of the next color to use in the plot.
-        color_idx = 0
-
-        for event in events:
-            if plot_idx >= num_plots:
-                break
-
-            feature_names = event.feature_names
-
-            xs = event.data[index].timestamps
-            uniform = is_uniform(xs)
-
-            plot_mask = np.full(len(xs), True)
-            if options.min_time is not None:
-                plot_mask = plot_mask & (xs >= options.min_time)
-            if options.max_time is not None:
-                plot_mask = plot_mask & (xs <= options.max_time)
-            if options.max_points is not None and len(xs) > options.max_points:
-                # Too many timestamps. Only keep the fist ones.
-                plot_mask = plot_mask & (
-                    np.cumsum(plot_mask) <= options.max_points
-                )
-
-            xs = xs[plot_mask]
-
-            if event.is_unix_timestamp:
-                # Matplotlib understands datetimes.
-                xs = [
-                    datetime.datetime.fromtimestamp(x, tz=datetime.timezone.utc)
-                    for x in xs
-                ]
-
-            if len(feature_names) == 0:
-                # There is not features to plot. Instead, plot the timestamps.
-                figs.append(
-                    _bokeh_sub_plot(
-                        xs=xs,
-                        ys=np.zeros(len(xs)),
-                        options=options,
-                        color=colors[color_idx % len(colors)],
-                        name="[sampling]",
-                        is_unix_timestamp=event.is_unix_timestamp,
-                        title=title,
-                        style=Style.vline,
-                    )
-                )
-                # Only print the index / title once
-                title = None
-
-                color_idx += 1
-                plot_idx += 1
-
-            for feature_idx, feature_name in enumerate(feature_names):
-                if plot_idx >= num_plots:
-                    # Too much plots are displayed already.
-                    break
-
-                ys = event.data[index].features[feature_idx][plot_mask]
-                if len(ys) == 0:
-                    all_ys_are_equal = True
-                else:
-                    all_ys_are_equal = np.all(ys == ys[0])
-
-                effective_stype = options.style
-                if effective_stype == Style.auto:
-                    if not uniform and (len(xs) <= 1000 or all_ys_are_equal):
-                        effective_stype = Style.marker
-                    else:
-                        effective_stype = Style.line
-
-                figs.append(
-                    _bokeh_sub_plot(
-                        xs=xs,
-                        ys=ys,
-                        options=options,
-                        color=colors[color_idx % len(colors)],
-                        name=feature_name,
-                        is_unix_timestamp=event.is_unix_timestamp,
-                        title=title,
-                        style=effective_stype,
-                    )
-                )
-
-                # Only print the index / title once
-                title = None
-
-                color_idx += 1
-                plot_idx += 1
-
-    if len(figs) > 1:
-        # Sync x-axes
-        js_vars = [f"p{fig_idx+1}_x_range" for fig_idx, fig in enumerate(figs)]
-        js_inputs = {}
-        core_code = ""
-        for js_var, fig in zip(js_vars, figs):
-            js_inputs[js_var] = fig.x_range
-
-            sub_core_code = "\n".join([f"""
-            {other_js_var}.start = start;
-            {other_js_var}.end = end;
-            """ for other_js_var in js_vars if other_js_var != js_var])
-
-            core_code += f"""
-            if (cb_obj == {js_var}) {{
-                const start =  {js_var}.start;
-                const end =  {js_var}.end;
-                {sub_core_code}
-            }}
-            """
-
-        callback = CustomJS(args=js_inputs, code=core_code)
-
-        for fig in figs:
-            fig.x_range.js_on_change("start", callback)
-            fig.x_range.js_on_change("end", callback)
-
-        figure_set = gridplot(
-            [[f] for f in figs],
-            merge_tools=True,
-            toolbar_location="right",
-            toolbar_options=dict(logo=None),
-        )
-    else:
-        figure_set = figs[0]
-        figure_set.toolbar.logo = None
-
-    output_notebook(hide_banner=True)
-    show(figure_set)
-    return figure_set
+BACKENDS = {"matplotlib": plot_matplotlib, "bokeh": plot_bokeh}
 
 
-def _bokeh_sub_plot(
-    xs,
-    ys,
-    options: Options,
-    color,
-    name: str,
-    is_unix_timestamp: bool,
-    title: Optional[str],
-    style: Style,
-) -> Any:
-    """Plots "(xs, ys)" on the axis "ax"."""
-
-    from bokeh.plotting import figure
-    from bokeh.io import output_notebook, show
-    from bokeh.layouts import gridplot, column
-    from bokeh.models import ColumnDataSource, CategoricalColorMapper, HoverTool
-    from bokeh.models import ColumnDataSource, CustomJS, Range1d
-    from bokeh.palettes import Dark2_5 as colors
-
-    tools = [
-        "xpan",
-        "pan",
-        "xwheel_zoom",
-        "ywheel_zoom",
-        "box_zoom",
-        "reset",
-        "undo",
-        "save",
-        "hover",
-    ]
-
-    fig_args = {}
-    if is_unix_timestamp:
-        fig_args["x_axis_type"] = "datetime"
-    if title:
-        fig_args["title"] = title
-
-    is_string = ys.dtype.type is np.str_ or ys.dtype.type is np.string_
-    if is_string:
-        unique_ys_values = list(set(ys))
-        fig_args["y_range"] = unique_ys_values
-    else:
-        unique_ys_values = None
-
-    fig = figure(
-        width=options.width_px,
-        height=options.height_per_plot_px,
-        tools=tools,
-        **fig_args,
+def error_message_import_backend(backend: str) -> str:
+    return (
+        f"Cannot plot with selected backend={backend}. Solutions: (1) Install"
+        f" {backend} e.g. 'pip install {backend}', or (2) use a different"
+        " plotting backen, for example with 'plot(..., backend=\"<other"
+        f' backend>"). The supported backends are: {list(BACKENDS.keys())}.'
     )
-
-    if options.min_time is not None or options.max_time is not None:
-        args = {}
-        if options.min_time is not None:
-            args["start"] = (
-                duration.convert_date_to_duration(options.min_time)
-                if not is_unix_timestamp
-                else options.min_time
-            )
-        if options.max_time is not None:
-            args["end"] = (
-                duration.convert_date_to_duration(options.max_time)
-                if not is_unix_timestamp
-                else options.max_time
-            )
-        fig.x_range = Range1d(**args)
-
-    data = {"x": xs, "y": ys, "color": color}
-
-    if is_string:
-        fig.circle(x=xs, y=ys)
-    elif style == Style.line:
-        fig.line(**data)
-    elif style == Style.marker:
-        fig.scatter(**data)
-    elif style == Style.vline:
-        fig.scatter(**data)
-    else:
-        raise ValueError("Non implemented style")
-
-    fig.yaxis.axis_label = name
-
-    return fig
-
-
-BACKENDS = {"matplotlib": _plot_matplotlib, "bokeh": _plot_bokeh}
