+++ temporian/implementation/numpy/data/plotter_matplotlib.py
@@ -1,0 +1,195 @@
+"""Matplotlib plotting backend."""
+
+import datetime
+from typing import Optional, List, Any
+
+import numpy as np
+
+from temporian.core.data import duration
+from temporian.implementation.numpy.data import plotter
+from temporian.implementation.numpy.data.event import NumpyEvent
+from temporian.implementation.numpy.data.plotter import (
+    Options,
+    Style,
+    is_uniform,
+    get_num_plots,
+)
+
+
+def plot_matplotlib(
+    events: List[NumpyEvent], indexes: List[tuple], options: Options
+):
+    import matplotlib.pyplot as plt
+    from matplotlib.cm import get_cmap
+
+    colors = get_cmap("tab10").colors
+
+    px = 1 / plt.rcParams["figure.dpi"]
+
+    num_plots = get_num_plots(events, indexes, options)
+
+    fig, axs = plt.subplots(
+        num_plots,
+        figsize=(
+            options.width_px * px,
+            options.height_per_plot_px * num_plots * px,
+        ),
+        squeeze=False,
+    )
+
+    # Actual plotting
+    plot_idx = 0
+    for index in indexes:
+        if plot_idx >= num_plots:
+            # Too much plots are displayed already.
+            break
+
+        # Note: Don't display the tuple parenthesis is the index contain a
+        # single dimension.
+        title = str(index[0] if len(index) == 1 else index)
+
+        # Index of the next color to use in the plot.
+        color_idx = 0
+
+        for event in events:
+            if plot_idx >= num_plots:
+                break
+
+            feature_names = event.feature_names
+
+            xs = event.data[index].timestamps
+            uniform = is_uniform(xs)
+
+            plot_mask = np.full(len(xs), True)
+            if options.min_time is not None:
+                plot_mask = plot_mask & (xs >= options.min_time)
+            if options.max_time is not None:
+                plot_mask = plot_mask & (xs <= options.max_time)
+            if options.max_points is not None and len(xs) > options.max_points:
+                # Too many timestamps. Only keep the fist ones.
+                plot_mask = plot_mask & (
+                    np.cumsum(plot_mask) <= options.max_points
+                )
+
+            xs = xs[plot_mask]
+
+            if event.is_unix_timestamp:
+                # Matplotlib understands datetimes.
+                xs = [
+                    datetime.datetime.fromtimestamp(x, tz=datetime.timezone.utc)
+                    for x in xs
+                ]
+
+            if len(feature_names) == 0:
+                # There is not features to plot. Instead, plot the timestamps.
+                _matplotlib_sub_plot(
+                    ax=axs[plot_idx, 0],
+                    xs=xs,
+                    ys=np.zeros(len(xs)),
+                    options=options,
+                    color=colors[color_idx % len(colors)],
+                    name="[sampling]",
+                    is_unix_timestamp=event.is_unix_timestamp,
+                    title=title,
+                    style=Style.vline,
+                )
+                # Only print the index / title once
+                title = None
+
+                color_idx += 1
+                plot_idx += 1
+
+            for feature_idx, feature_name in enumerate(feature_names):
+                if plot_idx >= num_plots:
+                    # Too much plots are displayed already.
+                    break
+
+                ys = event.data[index].features[feature_idx][plot_mask]
+                if len(ys) == 0:
+                    all_ys_are_equal = True
+                else:
+                    all_ys_are_equal = np.all(ys == ys[0])
+
+                effective_stype = options.style
+                if effective_stype == Style.auto:
+                    if not uniform and (len(xs) <= 1000 or all_ys_are_equal):
+                        effective_stype = Style.marker
+                    else:
+                        effective_stype = Style.line
+
+                _matplotlib_sub_plot(
+                    ax=axs[plot_idx, 0],
+                    xs=xs,
+                    ys=ys,
+                    options=options,
+                    color=colors[color_idx % len(colors)],
+                    name=feature_name,
+                    is_unix_timestamp=event.is_unix_timestamp,
+                    title=title,
+                    style=effective_stype,
+                )
+
+                # Only print the index / title once
+                title = None
+
+                color_idx += 1
+                plot_idx += 1
+
+    fig.tight_layout()
+    return fig
+
+
+def _matplotlib_sub_plot(
+    ax,
+    xs,
+    ys,
+    options: Options,
+    color,
+    name: str,
+    is_unix_timestamp: bool,
+    title: Optional[str],
+    style: Style,
+    **wargs,
+):
+    """Plots "(xs, ys)" on the axis "ax"."""
+
+    import matplotlib.ticker as ticker
+
+    if style == Style.line:
+        mat_style = {}  # Default
+    elif style == Style.marker:
+        mat_style = {"marker": "2", "linestyle": "None"}
+    elif style == Style.vline:
+        mat_style = {"marker": "|", "linestyle": "None"}
+    else:
+        raise ValueError("Non implemented style")
+
+    ax.plot(xs, ys, lw=0.5, color=color, **mat_style, **wargs)
+    if options.min_time is not None or options.max_time is not None:
+        args = {}
+        if options.min_time is not None:
+            args["left"] = (
+                duration.convert_date_to_duration(options.min_time)
+                if not is_unix_timestamp
+                else options.min_time
+            )
+        if options.max_time is not None:
+            args["right"] = (
+                duration.convert_date_to_duration(options.max_time)
+                if not is_unix_timestamp
+                else options.max_time
+            )
+        ax.set_xlim(**args)
+
+    ax.xaxis.set_tick_params(labelsize=8)
+    ax.xaxis.set_major_locator(ticker.MaxNLocator(10))
+    ax.xaxis.set_minor_locator(ticker.NullLocator())
+
+    ax.set_ylabel(name, size=8)
+    ax.yaxis.set_tick_params(labelsize=8)
+    ax.yaxis.set_major_locator(ticker.MaxNLocator(5))
+    ax.yaxis.set_minor_locator(ticker.NullLocator())
+
+    ax.grid(lw=0.4, ls="--", axis="x")
+    if title:
+        ax.set_title(title)
