+++ temporian/implementation/numpy/data/test/plotter_test.py
@@ -1,4 +1,4 @@
-from absl.testing import absltest
+from absl.testing import parameterized, absltest
 
 import numpy as np
 
@@ -7,14 +7,16 @@ from temporian.implementation.numpy.data.event import IndexData
 from temporian.implementation.numpy.data.event import NumpyEvent
 
 
-class PlotterTest(absltest.TestCase):
-    def test_plot(self):
+class PlotterTest(parameterized.TestCase):
+    @parameterized.parameters((True,), (False,))
+    def test_plot(self, interactive):
         event = NumpyEvent(
             data={
                 (1,): IndexData(
                     features=[
                         np.array([1, 2, 3]),
                         np.array([4, 5, 6]),
+                        np.array(["X", "Y", "X"]),
                     ],
                     timestamps=np.array([0.1, 0.2, 0.3]),
                 ),
@@ -22,19 +24,31 @@ class PlotterTest(absltest.TestCase):
                     features=[
                         np.array([7, 8]),
                         np.array([9, 10]),
+                        np.array(["X", "Z"]),
                     ],
                     timestamps=np.array([0.4, 0.5]),
                 ),
             },
-            feature_names=["a", "b"],
+            feature_names=["a", "b", "c"],
             index_names=["x"],
             is_unix_timestamp=False,
         )
 
-        _ = plotter.plot(event, indexes=None)
-        _ = plotter.plot(event, indexes=1)
-        _ = plotter.plot(event, indexes=[1, 2])
-        _ = plotter.plot(event, indexes=[(1,), (2,)])
+        _ = plotter.plot(
+            event, indexes=None, interactive=interactive, return_fig=True
+        )
+        _ = plotter.plot(
+            event, indexes=1, interactive=interactive, return_fig=True
+        )
+        _ = plotter.plot(
+            event, indexes=[1, 2], interactive=interactive, return_fig=True
+        )
+        _ = plotter.plot(
+            event,
+            indexes=[(1,), (2,)],
+            interactive=interactive,
+            return_fig=True,
+        )
 
     def test_is_uniform(self):
         self.assertTrue(plotter.is_uniform([]))
